api_yamdb/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
asgiref==3.2.10
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.0.4
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.1
PyJWT==2.1.0
psycopg2-binary==2.8.6
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0